addgoods

+++ administrator-side-software/src/main/java/com/example/administratorsidesoftware/service/WarehouseService.java
@@ -1,20 +1,15 @@
 package com.example.administratorsidesoftware.service;
 
 import cn.hutool.core.bean.BeanUtil;
-import com.baomidou.mybatisplus.core.metadata.TableInfo;
-import com.baomidou.mybatisplus.core.metadata.TableInfoHelper;
-import com.baomidou.mybatisplus.core.toolkit.Assert;
-import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.example.administratorsidesoftware.controller.DTO.WarehouseDTO;
+import com.example.administratorsidesoftware.entity.Position;
 import com.example.administratorsidesoftware.entity.Warehouse;
-import com.example.administratorsidesoftware.mapper.PositionMapper;
 import com.example.administratorsidesoftware.mapper.WarehouseMapper;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
-import java.io.Serializable;
-import java.util.Objects;
+import java.util.List;
 
 
 @Service
@@ -28,18 +23,25 @@ public class WarehouseService extends ServiceImpl<WarehouseMapper, Warehouse> {
     public boolean saveOrUpdateWarehouse(WarehouseDTO warehouseDTO) {
         Warehouse warehouse = new Warehouse();
         BeanUtil.copyProperties(warehouseDTO, warehouse);
-        if(warehouse.getWarehouseNo() != null){
-            if(warehouseMapper.selectById(warehouse.getWarehouseNo()) != null){
+        if (warehouse.getWarehouseNo() != null) {
+            if (warehouseMapper.selectById(warehouse.getWarehouseNo()) != null) {
                 return updateById(warehouse);
-            }else {
-               return save(warehouse) && positionService.saveWarehouse(warehouse);
+            } else {
+                return save(warehouse) && positionService.saveWarehouse(warehouse);
             }
         } else {
-           return save(warehouse) && positionService.saveWarehouse(warehouse);
+            return save(warehouse) && positionService.saveWarehouse(warehouse);
         }
     }
 
     public boolean removeWarehouse(Integer warehouseNo) {
-        return positionService.removeWarehouse(warehouseNo) && removeById(warehouseNo);
+
+        List<Position> positions = positionService.listNonEmptyPositionByWarehouse(warehouseNo);
+        //有东西的position列表
+        if (positions.isEmpty()) {
+            return positionService.removeWarehouse(warehouseNo) && removeById(warehouseNo);
+        } else {
+            return false;
+        }
     }
 }
